Implemented threading locks + merge sort fix
Previously we were using deep copies of the clients_online variable to avoid race conditions, effective but more of a patch rather than an actual fix (Plus it may have been the cause of some rare uncaught exceptions).
Now we are using thread locks to ensure the clients_online variable is accessed by a single thread at a time, preventing race conditions and no doubt making the application more stable as a whole.

Merge sort: Silly me, using min() to find the smallest element of a list when it is already sorted from smallest to largest, so just need to do list_name[0] :/